Generate site: `yo phaser-official:app`
I don't even know what to call the game yet, or the specifics of the theme nor
the exact mechanics.

At this stage, all I know is that there's a two-level map: the local action
in one island/sector/country/planet, where there's a zero-sum battle of forces;
and the high-level view of all of the islands/sectors/countries/planets
interacting with each other in a different way.